fix: redirect in JS after processing the auth link
A cookie set by the server in a 302 redirect response can sometimes not
be included in the follow-up request, particularly if the original
request was cross-site.
