+++ app/src/routes/auth/logout/+server.ts
@@ -1,6 +1,6 @@
 import { redirect, type RequestHandler } from '@sveltejs/kit';
 
-export const GET: RequestHandler = (event) => {
-	event.cookies.set('jwt', '', { path: '/', expires: new Date() });
+export const GET: RequestHandler = ({ cookies }) => {
+	cookies.set('jwt', '', { path: '/', expires: new Date() });
 	throw redirect(302, '/');
 };
